requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
reportlab==4.2.0
openai>=2.8.0
numpy==1.26.4
requests==2.32.3
pypdf==4.3.1
